Функция ресчета коэффициентов переведена на работу с типом long double

+++ gauss.c
@@ -1,37 +1,34 @@
-#include <stdio.h>
-#include <stdlib.h>
-#include <math.h>
-#include <string.h>
+#include "conmain.h"
 
 
-double * gauss(int M, double *t, double *f)
+long double * gauss(int M, double *t, double *f)
 {
     int size, i, j, v, k, d, p, J;
-    double max, ma[M][M+1], rotate[M+1], *X;
+    long double max, ma[M][M+1], rotate[M+1], *X;
 
-    X = (double *) malloc(M*sizeof(double));
+    X = (long double *) malloc(M*sizeof(long double));
 
     for (i = 0; i < M; i++){
         for (j = 0; j < M; j++){
-            ma[i][j] = pow(t[i], j);
+            ma[i][j] = powl((long double) t[i], j);
         }
-        ma[i][M] = f[i];
+        ma[i][M] = (long double) f[i];
     }
 
     for (i = 0; i < M; i++){
-        max = fabs(ma[i][i]);
+        max = fabsl(ma[i][i]);
         J = i;
         for (j = i+1; j < M; j++){
-            if (fabs(ma[j][i]) > max){
-                max = fabs(ma[j][i]);
+            if (fabsl(ma[j][i]) > max){
+                max = fabsl(ma[j][i]);
                 J = j;
             }
         }
 
         if (J != i){
-            memcpy(rotate, &ma[i][i], sizeof(double)*(M+1-i));
-            memcpy(&ma[i][i], &ma[J][i], sizeof(double)*(M+1-i));
-            memcpy(&ma[J][i], rotate, sizeof(double)*(M+1-i));
+            memcpy(rotate, &ma[i][i], sizeof(long double)*(M+1-i));
+            memcpy(&ma[i][i], &ma[J][i], sizeof(long double)*(M+1-i));
+            memcpy(&ma[J][i], rotate, sizeof(long double)*(M+1-i));
         }
 
         if (ma[i][i] != 0){
